+++ BookReader/plugins/plugin.vendor-fullscreen.js
@@ -2,15 +2,19 @@
  * Toggles browser's native fullscreen mode if available
  */
 
-
 (function(BR) {
-  var event_namespace = ".bookreader_vendor-fullscreen";
+  var event_namespace = '.bookreader_vendor-fullscreen';
+  var super_methods = {};
+  function isMobile() {
+    return (typeof window.orientation !== 'undefined') || (navigator.userAgent.indexOf('IEMobile') !== -1);
+  };
 
   jQuery.extend(BR.defaultOptions, {
     enableVendorFullscreenPlugin: true
   });
 
   BR.prototype.setup = (function(super_) {
+    super_methods.setup = super_;
     return function(options) {
       super_.call(this, options);
 
@@ -19,6 +23,7 @@
   })(BR.prototype.setup);
 
   BR.prototype.getInitialMode = (function(super_) {
+    super_methods.getInitialMode = super_;
     return function(params) {
       var nextMode = super_.call(this, params);
       if (this.isVendorFullscreenActive) {
@@ -29,6 +34,12 @@
   })(BR.prototype.getInitialMode);
 
   BR.prototype.init = (function(super_) {
+    if (isMobile()) {
+      for (var method in super_methods) {
+        BR.prototype[method] = super_methods[method];
+      }
+      return;
+    }
     return function() {
       super_.call(this);
 
@@ -138,8 +149,8 @@
       document.exitFullscreen();
     } else if (document.webkitExitFullscreen) {
       document.webkitExitFullscreen();
-    } else if (document.mozExitFullScreen) {
-      document.mozExitFullScreen();
+    } else if (document.mozCancelFullScreen) {
+      document.mozCancelFullScreen();
     } else if (document.msExitFullscreen) {
       document.msExitFullscreen();
     }
@@ -166,10 +177,11 @@
    * @see https://developer.mozilla.org/en-US/docs/Web/API/Document/fullscreenEnabled
    */
   BR.util.fullscreenAllowed = function() {
-    return document.fullscreenEnabled === true ||
-           document.webkitFullscreenEnabled === true ||
-           document.mozFullScreenEnabled === true ||
-           document.msFullScreenEnavled === true;
+    return (document.fullscreenEnabled ||
+           document.webkitFullscreenEnabled ||
+           document.mozFullScreenEnabled ||
+           document.msFullScreenEnabled) &&
+           !isMobile();
   };
 
   /**