+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 4.2.4
+- Skips initialization of vendor fullscreen plugin when on a mobile device
+
+# 4.2.3
+- Creates vendor native fullscreen plugin and example use page
+
 # 4.2.2
 - Adds options for flipSpeed and flipDelay, the latter used for autoplay